refactor: Rename API tab to DADOS and update icons
The 'API' tab has been renamed to 'DADOS' to better reflect its functionality of handling data related to the NFC-e. The associated icon has also been updated from 'Wifi' to 'Database' to align with the new tab name.

Additionally, the `SavedModel` type now includes an `updatedAt` field to track when a model was last modified. This change is part of a broader effort to improve data management and tracking within the application.